+++ src/codeactions/sortSymbols/alSortPropertiesCommand.ts
@@ -1,11 +1,11 @@
 import * as vscode from 'vscode';
-import { ALBaseSortCodeCommand } from "./alBaseSortCodeCommand";
 import { DevToolsExtensionContext } from "../../devToolsExtensionContext";
 import { AZSymbolInformation } from "../../symbollibraries/azSymbolInformation";
 import { AZSymbolKind } from '../../symbollibraries/azSymbolKind';
 import { AZDocumentSymbolsLibrary } from '../../symbollibraries/azDocumentSymbolsLibrary';
+import { ALCodeAction } from '../alCodeAction';
 
-export class ALSortPropertiesCommand extends ALBaseSortCodeCommand {
+export class ALSortPropertiesCommand extends ALCodeAction {
 
     constructor(context : DevToolsExtensionContext) {
         super(context, "SortProperties");
@@ -17,35 +17,19 @@ export class ALSortPropertiesCommand extends ALBaseSortCodeCommand {
 
             //collect list of objects in selection range
             if ((symbol) && (symbol.isALObject()) && (symbol.kind != AZSymbolKind.Interface) && 
+                (!symbol.containsDiagnostics) &&
                 (symbol.selectionRange) &&
-                (symbol.selectionRange.start.line == range.start.line))                
-                edit = this.prepareEdit(symbol, document, undefined);
+                (symbol.selectionRange.start.line == range.start.line)) {
 
-            if (edit) {
                 let action = new vscode.CodeAction("Sort properties", vscode.CodeActionKind.QuickFix);
-                action.edit = edit;
+                action.command = {
+                    command: "azALDevTools.sortProperties",
+                    title: "Sort Properties",
+                    arguments: [document, symbol.range]
+                }
                 actions.push(action);
             }
         }
-    }
-
-    protected prepareEdit(symbol: AZSymbolInformation, document: vscode.TextDocument, edit: vscode.WorkspaceEdit | undefined): vscode.WorkspaceEdit | undefined {
-        //collect PropertyList symbols
-        let propertyListSymbolsList: AZSymbolInformation[] = [];
-        symbol.collectChildSymbols(AZSymbolKind.PropertyList, true, propertyListSymbolsList);
- 
-        if (propertyListSymbolsList.length == 0)
-            return edit;
-
-        if (!edit)
-            edit = new vscode.WorkspaceEdit();
-        if (propertyListSymbolsList.length > 0) {
-            for (let i=0; i<propertyListSymbolsList.length; i++) {
-                if (!propertyListSymbolsList[i].containsDiagnostics)
-                    this.sortChildItems(document.uri, propertyListSymbolsList[i], AZSymbolKind.Property, edit);
-            }
-        }
-        return edit;
     }
 
 }