+++ src/codeactions/sortSymbols/alSortVariablesCommand.ts
@@ -1,15 +1,11 @@
 import * as vscode from 'vscode';
-import { ALBaseSortCodeCommand } from "./alBaseSortCodeCommand";
 import { DevToolsExtensionContext } from "../../devToolsExtensionContext";
 import { AZDocumentSymbolsLibrary } from "../../symbollibraries/azDocumentSymbolsLibrary";
 import { AZSymbolInformation } from "../../symbollibraries/azSymbolInformation";
 import { AZSymbolKind } from "../../symbollibraries/azSymbolKind";
+import { ALCodeAction } from '../alCodeAction';
 
-export class ALSortVariablesCommand extends ALBaseSortCodeCommand {
-    protected _typeByPriority: string[] = [
-        "record ", "report", "codeunit", "xmlport", "page", "query", "notification", 
-        "bigtext", "dateformula", "recordid", "recordref", "fieldref", "filterpagebuilder"];
-
+export class ALSortVariablesCommand extends ALCodeAction {
     constructor(context : DevToolsExtensionContext) {
         super(context, "SortVariables");
     }
@@ -21,133 +17,19 @@ export class ALSortVariablesCommand extends ALBaseSortCodeCommand {
             //collect list of objects in selection range
             if ((symbol) && 
                 ((symbol.isALObject()) || (symbol.kind == AZSymbolKind.VarSection) || (symbol.kind == AZSymbolKind.GlobalVarSection)) && 
+                (!symbol.containsDiagnostics) &&
                 (symbol.selectionRange) &&
-                (symbol.selectionRange.start.line == range.start.line))               
-                edit = this.prepareEdit(symbol, document, undefined);
+                (symbol.selectionRange.start.line == range.start.line)) {
 
-            if (edit) {
-                let action = new vscode.CodeAction("Sort variables", vscode.CodeActionKind.QuickFix);
-                action.edit = edit;
-                actions.push(action);
-            }
-        }
-    }
-
-    protected prepareEdit(symbol: AZSymbolInformation, document: vscode.TextDocument, edit: vscode.WorkspaceEdit | undefined): vscode.WorkspaceEdit | undefined {
-        if (symbol.containsDiagnostics)
-            return edit;
-        
-        //collect VariableSection symbols
-        let symbolsList: AZSymbolInformation[] = [];
-        if ((symbol.kind == AZSymbolKind.VarSection) || (symbol.kind == AZSymbolKind.GlobalVarSection))
-            symbolsList.push(symbol);
-        else
-            symbol.collectChildSymbolsByKindList([AZSymbolKind.VarSection,
-                AZSymbolKind.GlobalVarSection], true, symbolsList);
- 
-        if (symbolsList.length == 0)
-            return edit;
-
-        if (symbolsList.length > 0) {
-            for (let i=0; i<symbolsList.length; i++) {
-                edit = this.sortVariables(document, symbolsList[i], edit);
-            }
-        }
-        return edit;
-    }
-
-    protected sortVariables(document: vscode.TextDocument, symbol: AZSymbolInformation, editBuilder: vscode.WorkspaceEdit | undefined): vscode.WorkspaceEdit | undefined {
-        // Collect nodes
-        let childSymbolsList: AZSymbolInformation[] = [];
-        symbol.collectChildSymbolsByKindList([AZSymbolKind.VariableDeclaration, 
-            AZSymbolKind.VariableDeclarationName], false, childSymbolsList);
-        if (childSymbolsList.length == 0)
-            return editBuilder;
-
-        // Sort nodes
-        childSymbolsList.sort((symbolA, symbolB) => {
-            return this.compareSymbols(symbolA, symbolB);
-        });
-
-        // Produce the new sorted source
-        let newSource: string = "";
-
-        for (let i=0; i<childSymbolsList.length; i++) {
-            let childSymbol = childSymbolsList[i];
-            if (childSymbol.range) {
-                //get symbol range
-                let declRange = new vscode.Range(childSymbol.range.start.line, childSymbol.range.start.character, 
-                    childSymbol.range.end.line, childSymbol.range.end.character);
-            
-                //build new source code
-                newSource += document.getText(declRange);
-                if (childSymbol.kind == AZSymbolKind.VariableDeclarationName) {
-                    if ((i < (childSymbolsList.length - 1)) &&
-                        (childSymbolsList[i+1].kind == childSymbol.kind) &&
-                        (childSymbolsList[i+1].subtype == childSymbol.subtype)) {
-                        newSource += ', ';
-                        if (childSymbolsList[i+1].range!.start.line < childSymbolsList[i+1].range!.end.line)
-                            newSource += '\n';
-                    } else
-                        newSource += ': ' + childSymbol.subtype + ';\n';
+                    let action = new vscode.CodeAction("Sort variables", vscode.CodeActionKind.QuickFix);
+                    action.command = {
+                        command: "azALDevTools.sortVariables",
+                        title: "Sort Variables",
+                        arguments: [document, symbol.range]
+                    }
+                    actions.push(action);
                 }
-            }
         }
-        
-        // Delete the old unsorted columns and insert the new sorted source
-        if (symbol.contentRange) {
-            const deleteRange = new vscode.Range(symbol.contentRange.start.line, symbol.contentRange.start.character, 
-                symbol.contentRange.end.line, symbol.contentRange.end.character);
-
-            if (!editBuilder)
-                editBuilder = new vscode.WorkspaceEdit();
-
-            editBuilder.delete(document.uri, deleteRange);
-            editBuilder.insert(document.uri, deleteRange.start, newSource);
-        }
-
-        return editBuilder;
-    }
-
-    protected compareSymbols(symbolA: AZSymbolInformation, symbolB: AZSymbolInformation): number {
-        let subtypeA = this.preProcessTypeName(symbolA.elementsubtype);
-        let subtypeB = this.preProcessTypeName(symbolB.elementsubtype);
-        
-        let result: number = this.getTypePriority(subtypeA) - this.getTypePriority(subtypeB);        
-        if (result != 0)
-            return result;
-        
-        if ((subtypeA) && (subtypeB)) {
-            result = subtypeA.localeCompare(subtypeB, undefined, { numeric: true, sensitivity: 'base' });
-            if (result != 0)
-                return result;
-        }
-
-        subtypeA = this.preProcessTypeName(symbolA.subtype);
-        subtypeB = this.preProcessTypeName(symbolB.subtype);
-        if ((subtypeA) && (subtypeB)) {
-            result = subtypeA.localeCompare(subtypeB, undefined, { numeric: true, sensitivity: 'base' });
-            if (result != 0)
-                return result;
-        }
-
-        return symbolA.name.localeCompare(symbolB.name, undefined, { numeric: true, sensitivity: 'base' });
-    }
-
-    protected getTypePriority(typeName: string | undefined) {
-        if (typeName) {
-            typeName = typeName.toLowerCase();
-            for (let i=0; i<this._typeByPriority.length; i++)
-                if (typeName.startsWith(this._typeByPriority[i]))
-                    return i;
-        }
-        return this._typeByPriority.length;
-    }
-
-    protected preProcessTypeName(name: string | undefined) {
-        if (!name)
-            return name;
-        return name.replace(/"/g, '');
     }
 
 }